refactor: split iac/utils
This is required to fix a circular dependency when ScanMode and
ScanContext are moved in one of the next commits.
